재고현황,제품상세수정 구현_20220517

+++ Convenience_POS_System/src/main/java/com/example/convenience_pos_system/dao/ProductHistoryDao.java
@@ -50,7 +50,7 @@ public class ProductHistoryDao {
         productHistory.setId(keyValue.longValue());
     }
 
-    public List<ProductHistory> selectByPid(int pid) {
+    public List<ProductHistory> selectByPid(Long pid) {
         List<ProductHistory> results = jdbcTemplate.query(
                 "select * from PRODUCT_HISTORY_TB where PID = ?",
                 new RowMapper<ProductHistory>() {
